requirements.txt
uvicorn
fastapi
pydantic
scikit-learn
numpy==1.24.3
pandas
pillow